+++ apps/chronicles/mad.config.ts
@@ -1,24 +1,34 @@
-import { MadConfig } from "@equinor/mad-core";
+import {MadConfig} from "@equinor/mad-core";
 import Logo from "./assets/images/icon.png";
 import { ImageSourcePropType } from "react-native";
+import { getBuildNumber } from "./settings";
 
 export const config: MadConfig = {
     appVersion: "1.0.0",
     servicePortalName: "Chronicles",
+    currentEnvironment: "prod",
     serviceNowConfigurationItem: "MAD",
-    environment: "prod",
     language: {
         supportedLanguages: [
             { code: "en", name: "English" },
             { code: "nb", name: "Norwegian" },
             { code: "pt", name: "Portuguese" },
         ],
+        skipOnboarding: false,
     },
     authentication: {
-        redirectUri: "msauth.com.equinor.mad.chronicles://auth",
-        redirectUriWeb: "http://localhost:8081",
-        clientId: "49222fe1-4e0a-4310-9e81-1a2c3eb9b2ed",
-        scopes: ["0a429637-3fe1-4452-bd95-c87923ba340b/user_impersonation"],
+        prod: {
+            redirectUri: "msauth.com.equinor.mad.chronicles://auth",
+            redirectUriWeb: "http://localhost:8081",
+            clientId: "49222fe1-4e0a-4310-9e81-1a2c3eb9b2ed",
+            scopes: ["0a429637-3fe1-4452-bd95-c87923ba340b/user_impersonation"],
+        },
+        test: {
+            redirectUri: "msauth.com.equinor.mad.chronicles://auth",
+            redirectUriWeb: "http://localhost:8081",
+            clientId: "49222fe1-4e0a-4310-9e81-1a2c3eb9b2ed",
+            scopes: ["830a7388-cd89-4e25-a631-bd615bf225a4/user_impersonation"],
+        },
     },
     login: {
         title: "Chronicles",
@@ -33,6 +43,6 @@ export const config: MadConfig = {
     },
     about: {
         endpoints: [],
-        buildNumber: "",
+        buildNumber: getBuildNumber(),
     },
 };