+++ packages/components/src/components/Popover/Popover.tsx
@@ -12,7 +12,7 @@ export type PopoverProps = {
     /**
      * A reference to the element that the popover should appear around.
      */
-    anchorEl: React.MutableRefObject<View | null>;
+    anchorEl: React.RefObject<View>;
     /**
      * A boolean value indicating whether or not the popover should be open.
      */