+++ packages/components/src/components/ProgressIndicator/useNoProgressAnimation.ts
@@ -1,37 +1,30 @@
 import { useEffect, useRef } from "react";
-import { Animated } from "react-native";
-import { useToken } from "../../hooks/useToken";
+import { Platform, Animated } from "react-native";
 
-/**
- * Creates a smooth looping animated value.
- * @param value An initial value to start the loop around.
- * @returns An animated value between 0 and 1 representing the current loop progress.
- */
-export const useNoProgressAnimation = (value?: number) => {
-    const loopValue = useRef(new Animated.Value(value ?? 0)).current;
-    const token = useToken();
+export const useNoProgressAnimation = (value = 0) => {
+    const loopValue = useRef(new Animated.Value(value)).current;
 
-    const endlessAnimation = Animated.loop(
-        Animated.timing(loopValue, {
-            toValue: 1,
-            duration: 1500,
-            useNativeDriver: true,
-        }),
-    );
-
-    const resetAnimation = Animated.timing(loopValue, {
-        toValue: 0,
-        duration: token.timing.animation.normal,
+    const animate = Animated.timing(loopValue, {
+        toValue: 1,
+        duration: 1500,
         useNativeDriver: true,
     });
 
+    const startAnimation = () => {
+        animate.start(() => {
+            loopValue.setValue(0);
+            startAnimation();
+        });
+    };
+
     useEffect(() => {
-        if (value !== undefined) {
-            resetAnimation.start(() => endlessAnimation.stop());
+        if (Platform.OS === 'web') {
+            startAnimation(); 
         } else {
-            endlessAnimation.start();
+            Animated.loop(animate).start();
         }
-        // eslint-disable-next-line react-hooks/exhaustive-deps -- adding animations to deps cause infinite loop glitch
+        return () => loopValue.stopAnimation();
+
     }, [value]);
 
     return loopValue;