+++ packages/components/src/hooks/useDynamicStyleSheet.ts
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useMemo, DependencyList } from "react";
 import { StyleSheet } from "react-native";
 
 const createStyleSheet = <T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<unknown>>(
@@ -15,7 +15,7 @@ export const useDynamicStyleSheet = <
     T extends StyleSheet.NamedStyles<T> | StyleSheet.NamedStyles<unknown>,
 >(
     style: () => T | StyleSheet.NamedStyles<T>,
-    dependencies?: React.DependencyList,
+    dependencies: DependencyList,
 ) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps -- we only want to update when deps change
     return useMemo(() => createStyleSheet(style()), dependencies);