+++ packages/components/src/index.ts
@@ -20,6 +20,7 @@ export { alert } from "./components/Dialog";
 export * from "./components/Environment";
 export * from "./components/ErrorBoundary";
 export * from "./components/SelectionControls";
+export * from "./components/Chip";
 
 export * from "./styling/EDSStyleSheet";
 export * from "./styling/types";