+++ packages/core/src/components/service-message/utils/resolveMessageFromServiceMessage.ts
@@ -1,0 +1,16 @@
+import { ServiceMessage } from "../types";
+
+export const resolveMessageFromServiceMessage = (
+    serviceMessage: ServiceMessage,
+    isError: boolean,
+): { message: string; url?: string | null } => {
+    if (isError)
+        return {
+            message: "Could not retrieve service message.",
+        };
+
+    return {
+        message: serviceMessage.message,
+        url: serviceMessage.urlString,
+    };
+};