+++ packages/core/src/store/index.ts
@@ -1,0 +1,3 @@
+export * from "./demo-mode";
+export * from "./language";
+export {useMadConfig, setEnvironment, getConfig } from "./mad-config";