+++ packages/skia-draw/src/premades/SignaturePad/SignaturePad.tsx
@@ -1,12 +1,17 @@
 import React, { forwardRef, useImperativeHandle, useRef } from "react";
-import { Canvas } from "../../Canvas";
-import { SignaturePadProps, SkiaDrawHandle, SnapshotHandle } from "../../types";
+import { Canvas } from "../../Canvas/Canvas";
+import { SignaturePadProps } from "../../types";
 import { View, StyleSheet, Text } from "react-native";
 import { SkRect } from "@shopify/react-native-skia";
 import { Button } from "@equinor/mad-components";
+import {
+    CanvasControlProvider,
+    CanvasControls,
+    CanvasImageControls,
+} from "../../CanvasControlProvider";
 
-export const SignaturePad = forwardRef<SnapshotHandle, SignaturePadProps>((props, ref) => {
-    const canvasRef = useRef<SkiaDrawHandle>(null);
+export const SignaturePad = forwardRef<CanvasImageControls, SignaturePadProps>((props, ref) => {
+    const canvasRef = useRef<CanvasControls>(null);
     useImperativeHandle(ref, () => ({
         makeImageSnapshot: (rect?: SkRect) =>
             canvasRef.current?.makeImageSnapshot(rect) ?? undefined,
@@ -14,28 +19,28 @@ export const SignaturePad = forwardRef<SnapshotHandle, SignaturePadProps>((props
     const canvasStyle = StyleSheet.flatten([styles.canvas, { height: props.height ?? 200 }]);
 
     return (
-        <View style={styles.container}>
-            <View style={styles.canvasContainer}>
-                <Canvas
-                    ref={canvasRef}
-                    initialDrawColor={"black"}
-                    initialStrokeWidth={3}
-                    style={canvasStyle}
-                    onLayout={props.onLayout}
-                >
-                    {props.children}
-                </Canvas>
-                {props.withLabel && <Text style={styles.signatureLabel}>Signature</Text>}
+        <CanvasControlProvider
+            canvasRef={canvasRef}
+            initialStrokeWeight={3}
+            initialToolColor={"black"}
+        >
+            <View style={styles.container}>
+                <View style={styles.canvasContainer}>
+                    <Canvas ref={canvasRef} style={canvasStyle} onLayout={props.onLayout}>
+                        {props.children}
+                    </Canvas>
+                    {props.withLabel && <Text style={styles.signatureLabel}>Signature</Text>}
+                </View>
+                <View style={styles.clearSignatureButtonContainer}>
+                    <Button
+                        variant="outlined"
+                        title="Clear signature"
+                        color="secondary"
+                        onPress={() => canvasRef.current?.clear()}
+                    />
+                </View>
             </View>
-            <View style={styles.clearSignatureButtonContainer}>
-                <Button
-                    variant="outlined"
-                    title="Clear signature"
-                    color="secondary"
-                    onPress={() => canvasRef.current?.clear()}
-                />
-            </View>
-        </View>
+        </CanvasControlProvider>
     );
 });