+++ packages/skia-draw/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @equinor/react-native-skia-draw
 
+## 0.3.2
+
+### Patch Changes
+
+-   Updated dependencies [97dca15]
+-   Updated dependencies [97dca15]
+    -   @equinor/mad-components@0.8.4
+
+## 0.3.1
+
+### Patch Changes
+
+-   Updated dependencies [59860d3]
+-   Updated dependencies [6f62312]
+    -   @equinor/mad-components@0.8.3
+
+## 0.3.0
+
+### Minor Changes
+
+-   e7cc7e0: **Breaking change**: `SnapShotHandle` type is now replaced with `CanvasImageControls`
+    type.
+-   e7cc7e0: Reworked the canvas control system. The new system allows for much more fine control
+    through your custom control panel. See the README for more information
+
+### Patch Changes
+
+-   Updated dependencies [38e818d]
+-   Updated dependencies [cf131da]
+-   Updated dependencies [e941e2e]
+    -   @equinor/mad-components@0.8.2
+
 ## 0.2.27
 
 ### Patch Changes